refactor: move DB startup logic to internal package
