+++ internal/app/run.go
@@ -7,8 +7,6 @@ import (
 	"log"
 	"net/http"
 	"os"
-	"path/filepath"
-	"strings"
 	"time"
 
 	goa4web "github.com/arran4/goa4web"
@@ -18,6 +16,7 @@ import (
 	corelanguage "github.com/arran4/goa4web/core/language"
 	adminhandlers "github.com/arran4/goa4web/handlers/admin"
 	userhandlers "github.com/arran4/goa4web/handlers/user"
+	dbstart "github.com/arran4/goa4web/internal/dbstart"
 	email "github.com/arran4/goa4web/internal/email"
 	emailutil "github.com/arran4/goa4web/internal/emailutil"
 	middleware "github.com/arran4/goa4web/internal/middleware"
@@ -60,11 +59,11 @@ func RunWithConfig(ctx context.Context, cfg runtimeconfig.RuntimeConfig, session
 	}
 	common.Version = version
 
-	if err := performStartupChecks(cfg); err != nil {
+	if err := dbstart.PerformStartupChecks(cfg); err != nil {
 		return fmt.Errorf("startup checks: %w", err)
 	}
 
-	dbPool := goa4web.GetDBPool()
+	dbPool := dbstart.GetDBPool()
 	if err := corelanguage.ValidateDefaultLanguage(context.Background(), goa4web.New(dbPool), cfg.DefaultLanguage); err != nil {
 		return fmt.Errorf("default language: %w", err)
 	}
@@ -144,33 +143,3 @@ type routerWrapper interface {
 type routerWrapperFunc func(http.Handler) http.Handler
 
 func (f routerWrapperFunc) Wrap(h http.Handler) http.Handler { return f(h) }
-
-func performStartupChecks(cfg runtimeconfig.RuntimeConfig) error {
-	if ue := goa4web.InitDB(cfg); ue != nil {
-		return fmt.Errorf("%s: %w", ue.ErrorMessage, ue.Err)
-	}
-	if err := checkUploadDir(cfg); err != nil {
-		return err
-	}
-	return nil
-}
-
-func checkUploadDir(cfg runtimeconfig.RuntimeConfig) error {
-	if cfg.ImageUploadDir == "" {
-		return fmt.Errorf("image upload directory not set")
-	}
-	if strings.HasPrefix(cfg.ImageUploadDir, "s3://") {
-		// TODO: validate S3 upload targets
-		return nil
-	}
-	info, err := os.Stat(cfg.ImageUploadDir)
-	if err != nil || !info.IsDir() {
-		return fmt.Errorf("image upload directory invalid: %w", err)
-	}
-	test := filepath.Join(cfg.ImageUploadDir, ".check")
-	if err := os.WriteFile(test, []byte("ok"), 0644); err != nil {
-		return fmt.Errorf("image upload directory not writable: %w", err)
-	}
-	os.Remove(test)
-	return nil
-}
